+++ src/core.py
@@ -141,6 +141,7 @@ def forward(rundir_or_config_path, event_names):
 def harvest(rundir, problem=None, nbest=10, force=False, weed=0):
 
     env = Environment([rundir])
+    optimiser = env.get_optimiser()
     nchains = env.get_optimiser().nchains
 
     if problem is None:
@@ -152,7 +153,6 @@ def harvest(rundir, problem=None, nbest=10, force=False, weed=0):
 
     logger.info('Harvesting problem "%s"...' % problem.name)
 
-    optimiser = load_optimiser_info(rundir)
     dumpdir = op.join(rundir, 'harvest')
     if op.exists(dumpdir):
         if force:
@@ -164,7 +164,7 @@ def harvest(rundir, problem=None, nbest=10, force=False, weed=0):
 
     ibests_list = []
     ibests = []
-    gms = problem.combine_misfits(misfits)
+    gms = bootstrap_misfits[:, 0]
     isort = num.argsort(gms)
 
     ibests_list.append(isort[:nbest])
@@ -442,10 +442,10 @@ g_state = {}
 
 def go(environment,
        force=False, preserve=False,
-       nparallel=1, status='state'):
+       nparallel=1, status='state', nthreads=0):
 
     g_data = (environment, force, preserve,
-              status, nparallel)
+              status, nparallel, nthreads)
     g_state[id(g_data)] = g_data
 
     nevents = environment.nevents_selected
@@ -460,7 +460,7 @@ def go(environment,
 
 def process_event(ievent, g_data_id):
 
-    environment, force, preserve, status, nparallel = \
+    environment, force, preserve, status, nparallel, nthreads = \
         g_state[g_data_id]
 
     config = environment.get_config()
@@ -513,6 +513,8 @@ def process_event(ievent, g_data_id):
     config.change_basepath(basepath)
 
     optimiser = config.optimiser_config.get_optimiser()
+    optimiser.set_nthreads(nthreads)
+
     optimiser.init_bootstraps(problem)
     problem.dump_problem_info(rundir)
 
@@ -582,8 +584,7 @@ class ResultStats(Object):
     parameter_stats_list = List.T(ParameterStats.T())
 
 
-def make_stats(problem, xs, misfits, pnames=None):
-    gms = problem.combine_misfits(misfits)
+def make_stats(problem, models, gms, pnames=None):
     ibest = num.argmin(gms)
     rs = ResultStats(problem=problem)
     if pnames is None:
@@ -591,7 +592,7 @@ def make_stats(problem, xs, misfits, pnames=None):
 
     for pname in pnames:
         iparam = problem.name_to_index(pname)
-        vs = problem.extract(xs, iparam)
+        vs = problem.extract(models, iparam)
         mi, p5, p16, median, p84, p95, ma = map(float, num.percentile(
             vs, [0., 5., 16., 50., 84., 95., 100.]))
 
@@ -676,9 +677,12 @@ def export(what, rundirs, type=None, pnames=None, filename=None):
 
     header = None
     for rundir in rundirs:
-        problem, xs, misfits, bootstrap_misfits, _ = \
-            load_problem_info_and_data(
-                rundir, subset='harvest')
+        env = Environment(rundir)
+        history = env.get_history(subset='harvest')
+
+        problem = history.problem
+        models = history.models
+        misfits = history.get_primary_chain_misfits()
 
         if type == 'vector':
             pnames_take = pnames_clean or \
@@ -703,21 +707,25 @@ def export(what, rundirs, type=None, pnames=None, filename=None):
             indices = None
 
         if what == 'best':
-            x_best, gm_best = stats.get_best_x_and_gm(problem, xs, misfits)
+            x_best = history.get_best_model()
+            gm_best = history.get_best_misfit()
             dump(x_best, gm_best, indices)
 
         elif what == 'mean':
-            x_mean, gm_mean = stats.get_mean_x_and_gm(problem, xs, misfits)
+            x_mean = history.get_mean_model()
+            gm_mean = history.get_mean_misfit(chain=0)
             dump(x_mean, gm_mean, indices)
 
         elif what == 'ensemble':
-            gms = problem.combine_misfits(misfits)
+            gms = misfits
             isort = num.argsort(gms)
             for i in isort:
-                dump(xs[i], gms[i], indices)
+                dump(models[i], gms[i], indices)
 
         elif what == 'stats':
-            rs = make_stats(problem, xs, misfits, pnames_clean)
+            rs = make_stats(problem, models,
+                            history.get_primary_chain_misfits(),
+                            pnames_clean)
             if shortform:
                 print(' ', format_stats(rs, pnames), file=out)
             else: