+++ src/problems/base.py
@@ -25,7 +25,6 @@ from grond import stats
 
 from grond.version import __version__
 
-
 guts_prefix = 'grond'
 logger = logging.getLogger('grond.problems.base')
 km = 1e3
@@ -36,6 +35,21 @@ g_rstate = num.random.RandomState()
 
 def nextpow2(i):
     return 2**int(math.ceil(math.log(i)/math.log(2.)))
+
+
+def correlated_weights(values, weight_matrix):
+    '''
+    Applies correlated weights to values
+
+    The resulting weighed values have to be squared! Check out
+    :meth:`Problem.combine_misfits` for more information.
+
+    :param values: Misfits or norms as :class:`numpy.Array`
+    :param weight: Weight matrix, commonly the inverse of covariance matrix
+
+    :returns: :class:`numpy.Array` weighted values
+    '''
+    return num.matmul(values, weight_matrix)
 
 
 class ProblemConfig(Object):
@@ -137,7 +151,7 @@ class Problem(Object):
         guts.dump(self, filename=fn)
 
     def dump_problem_data(
-            self, dirname, x, misfits, bootstraps=None,
+            self, dirname, x, misfits, chains=None,
             sampler_context=None):
 
         fn = op.join(dirname, 'models')
@@ -150,10 +164,10 @@ class Problem(Object):
         with open(fn, 'ab') as f:
             misfits.astype('<f8').tofile(f)
 
-        if bootstraps is not None:
-            fn = op.join(dirname, 'bootstraps')
+        if chains is not None:
+            fn = op.join(dirname, 'chains')
             with open(fn, 'ab') as f:
-                bootstraps.astype('<f8').tofile(f)
+                chains.astype('<f8').tofile(f)
 
         if sampler_context is not None:
             fn = op.join(dirname, 'choices')
@@ -278,8 +292,7 @@ class Problem(Object):
     def get_target_weights(self):
         if self._target_weights is None:
             self._target_weights = num.concatenate(
-                [target.get_combined_weight()
-                 for target in self.targets])
+               [target.get_combined_weight() for target in self.targets])
 
         return self._target_weights
 
@@ -305,13 +318,14 @@ class Problem(Object):
         '''
 
         exp, root = self.get_norm_functions()
-
         family, nfamilies = self.get_family_mask()
+
         ws = num.zeros(ns.shape)
         for ifamily in range(nfamilies):
             mask = family == ifamily
             ws[:, mask] = (1.0 / root(
                 num.nansum(exp(ns[:, mask]), axis=1)))[:, num.newaxis]
+
         return ws
 
     def get_reference_model(self, expand=False):
@@ -367,6 +381,7 @@ class Problem(Object):
             self, misfits,
             extra_weights=None,
             extra_residuals=None,
+            extra_correlated_weights=dict(),
             get_contributions=False):
 
         '''
@@ -388,6 +403,12 @@ class Problem(Object):
             to the residuals, indexed as
             ``extra_residuals[ibootstrap, iresidual]``.
 
+        :param extra_correlated_weights: if a dictionary of
+            ``imisfit: correlated weight matrix`` is passed a correlated
+            weight matrix is applied to the misfit and normalisation values.
+            `imisfit` is the starting index in the misfits vector the
+            correlated weight matrix applies to.
+
         :param get_contributions: get the weighted and perturbed contributions
             (don't do the sum).
 
@@ -397,62 +418,83 @@ class Problem(Object):
             ``misfits[imodel, ibootstrap]`` with the misfit for every model and
             weighting/residual set is returned.
         '''
-
-        exp, root = self.get_norm_functions()
-
         if misfits.ndim == 2:
             misfits = misfits[num.newaxis, :, :]
             return self.combine_misfits(
-                misfits,
-                extra_weights,
-                extra_residuals,
-                get_contributions)[0, ...]
+                misfits, extra_weights, extra_residuals,
+                extra_correlated_weights, get_contributions)[0, ...]
+
+        if extra_weights is None and extra_residuals is None:
+            return self.combine_misfits(
+                misfits, False, False,
+                extra_correlated_weights, get_contributions)[:, 0]
 
         assert misfits.ndim == 3
-        assert extra_weights is None or extra_weights.ndim == 2
-        assert extra_residuals is None or extra_residuals.ndim == 2
+        assert not num.any(extra_weights) or extra_weights.ndim == 2
+        assert not num.any(extra_residuals) or extra_residuals.ndim == 2
 
-        if extra_weights is not None or extra_residuals is not None:
-            if extra_weights is not None:
-                w = extra_weights[num.newaxis, :, :] \
-                    * self.get_target_weights()[num.newaxis, num.newaxis, :] \
-                    * self.inter_family_weights2(
-                        misfits[:, :, 1])[:, num.newaxis, :]
-            else:
-                w = 1.0
+        if self.norm_exponent != 2 and extra_correlated_weights:
+            raise GrondError('Correlated weights can only be used '
+                             ' with norm_exponent=2')
 
-            if extra_residuals is not None:
-                r = extra_residuals[num.newaxis, :, :]
-            else:
-                r = 0.0
+        exp, root = self.get_norm_functions()
 
-            if get_contributions:
-                return exp(w*(misfits[:, num.newaxis, :, 0]+r)) \
-                    / num.nansum(
-                        exp(w*misfits[:, num.newaxis, :, 1]),
-                        axis=2)[:, :, num.newaxis]
+        nmodels = misfits.shape[0]
+        nmisfits = misfits.shape[1]  # noqa
 
-            res = root(
-                num.nansum(exp(w*(misfits[:, num.newaxis, :, 0]+r)), axis=2) /
-                num.nansum(exp(w*(misfits[:, num.newaxis, :, 1])), axis=2))
-            assert res[res < 0].size == 0
-            return res
-        else:
-            w = self.get_target_weights()[num.newaxis, :] \
-                * self.inter_family_weights2(misfits[:, :, 1])
+        mf = misfits[:, num.newaxis, :, :].copy()
 
-            r = self.get_target_weights()[num.newaxis, :] \
-                * self.inter_family_weights2(misfits[:, :, 1])
+        if num.any(extra_residuals):
+            mf = mf + extra_residuals[num.newaxis, :, :, num.newaxis]
 
-            if get_contributions:
-                return exp(w*misfits[:, :, 0]) \
-                    / num.nansum(
-                        exp(w*misfits[:, :, 1]),
-                        axis=1)[:, num.newaxis]
+        res = mf[..., 0]
+        norms = mf[..., 1]
 
-            return root(
-                num.nansum(exp(w*misfits[:, :, 0]), axis=1) /
-                num.nansum(exp(w*misfits[:, :, 1]), axis=1))
+        for imisfit, corr_weight_mat in extra_correlated_weights.items():
+
+            jmisfit = imisfit + corr_weight_mat.shape[0]
+
+            for imodel in range(nmodels):
+                corr_res = res[imodel, :, imisfit:jmisfit]
+                corr_norms = norms[imodel, :, imisfit:jmisfit]
+
+                res[imodel, :, imisfit:jmisfit] = \
+                    correlated_weights(corr_res, corr_weight_mat)
+
+                norms[imodel, :, imisfit:jmisfit] = \
+                    correlated_weights(corr_norms, corr_weight_mat)
+
+        # Apply normalization family weights (these weights depend on
+        # on just calculated correlated norms!)
+        weights_fam = \
+            self.inter_family_weights2(norms[:, 0, :])[:, num.newaxis, :]
+
+        weights_fam = exp(weights_fam)
+
+        res = exp(res)
+        norms = exp(norms)
+
+        res *= weights_fam
+        norms *= weights_fam
+
+        weights_tar = self.get_target_weights()[num.newaxis, num.newaxis, :]
+        if num.any(extra_weights):
+            weights_tar = weights_tar * extra_weights[num.newaxis, :, :]
+
+        weights_tar = exp(weights_tar)
+
+        res = res * weights_tar
+        norms = norms * weights_tar
+
+        if get_contributions:
+            return res / num.nansum(norms, axis=2)[:, :, num.newaxis]
+
+        result = root(
+            num.nansum(res, axis=2) /
+            num.nansum(norms, axis=2))
+
+        assert result[result < 0].size == 0
+        return result
 
     def make_family_mask(self):
         family_names = set()
@@ -622,9 +664,12 @@ class ModelHistory(object):
         self._bootstraps_buffer = None
         self._sample_contexts_buffer = None
 
+        self._sorted_misfit_idx = {}
+
         self.models = None
         self.misfits = None
         self.bootstrap_misfits = None
+
         self.sampler_contexts = None
 
         self.nmodels_capacity = self.nmodels_capacity_min
@@ -637,7 +682,7 @@ class ModelHistory(object):
 
     @staticmethod
     def verify_rundir(rundir):
-        _rundir_files = ['misfits', 'models']
+        _rundir_files = ('misfits', 'models')
 
         if not op.exists(rundir):
             raise ProblemDataNotAvailable(
@@ -731,6 +776,7 @@ class ModelHistory(object):
                 self._bootstraps_buffer = bootstraps_buffer
 
     def clear(self):
+        assert self.mode != 'r', 'History is read-only, cannot clear.'
         self.nmodels = 0
         self.nmodels_capacity = self.nmodels_capacity_min
 
@@ -771,6 +817,8 @@ class ModelHistory(object):
                     if bootstrap_misfits is not None else None,
                     sampler_contexts[i, :]
                     if sampler_contexts is not None else None)
+
+        self._sorted_misfit_idx.clear()
 
         self.emit('extend', nmodels, n, models, misfits, sampler_contexts)
 
@@ -820,6 +868,11 @@ class ModelHistory(object):
             new_sampler_contexts)
 
     def add_listener(self, listener):
+        ''' Add a listener to the history
+
+        The listening class can implement the following methods:
+        * ``extend``
+        '''
         self.listeners.append(listener)
 
     def emit(self, event_name, *args, **kwargs):
@@ -922,6 +975,51 @@ class ModelHistory(object):
             for (icluster, percentage, models)
             in self.models_by_cluster(cluster_attribute)]
 
+    def get_sorted_misfits_idx(self, chain=0):
+        if chain not in self._sorted_misfit_idx.keys():
+            self._sorted_misfit_idx[chain] = num.argsort(
+                self.bootstrap_misfits[:, chain])
+
+        return self._sorted_misfit_idx[chain]
+
+    def get_sorted_misfits(self, chain=0):
+        isort = self.get_sorted_misfits_idx(chain)
+        return self.bootstrap_misfits[:, chain][isort]
+
+    def get_sorted_models(self, chain=0):
+        isort = self.get_sorted_misfits_idx(chain=0)
+        return self.models[isort, :]
+
+    def get_sorted_primary_misfits(self):
+        return self.get_sorted_misfits(chain=0)
+
+    def get_sorted_primary_models(self):
+        return self.get_sorted_models(chain=0)
+
+    def get_best_model(self, chain=0):
+        return self.get_sorted_models(chain)[0, ...]
+
+    def get_best_misfit(self, chain=0):
+        return self.get_sorted_misfits(chain)[0]
+
+    def get_mean_model(self):
+        return num.mean(self.models, axis=0)
+
+    def get_mean_misfit(self, chain=0):
+        return num.mean(self.bootstrap_misfits[:, chain])
+
+    def get_best_source(self, chain=0):
+        return self.problem.get_source(self.get_best_model(chain))
+
+    def get_mean_source(self, chain=0):
+        return self.problem.get_source(self.get_mean_model())
+
+    def get_chain_misfits(self, chain=0):
+        return self.bootstrap_misfits[:, chain]
+
+    def get_primary_chain_misfits(self):
+        return self.get_chain_misfits(chain=0)
+
 
 def get_nmodels(dirname, problem):
     fn = op.join(dirname, 'models')
@@ -959,6 +1057,13 @@ def load_problem_info(dirname):
 
 def load_problem_data(dirname, problem, nmodels_skip=0, nchains=None):
 
+    def get_chains_fn():
+        for fn in (op.join(dirname, 'bootstraps'),
+                   op.join(dirname, 'chains')):
+            if op.exists(fn):
+                return fn
+        return False
+
     try:
         nmodels = get_nmodels(dirname, problem) - nmodels_skip
 
@@ -981,17 +1086,17 @@ def load_problem_data(dirname, problem, nmodels_skip=0, nchains=None):
                 .astype(num.float)
         misfits = misfits.reshape((nmodels, problem.nmisfits, 2))
 
-        bootstraps = None
-        fn = op.join(dirname, 'bootstraps')
-        if op.exists(fn) and nchains is not None:
+        chains = None
+        fn = get_chains_fn()
+        if fn and nchains is not None:
             with open(fn, 'r') as f:
                 f.seek(nmodels_skip * nchains * 8)
-                bootstraps = num.fromfile(
+                chains = num.fromfile(
                         f, dtype='<f8',
                         count=nmodels*nchains)\
                     .astype(num.float)
 
-            bootstraps = bootstraps.reshape((nmodels, nchains))
+            chains = chains.reshape((nmodels, nchains))
 
         sampler_contexts = None
         fn = op.join(dirname, 'choices')
@@ -1009,7 +1114,7 @@ def load_problem_data(dirname, problem, nmodels_skip=0, nchains=None):
         raise ProblemDataNotAvailable(
             'No problem data available (%s).' % dirname)
 
-    return models, misfits, bootstraps, sampler_contexts
+    return models, misfits, chains, sampler_contexts
 
 
 __all__ = '''