+++ src/targets/waveform/plot.py
@@ -386,9 +386,11 @@ class FitsWaveformEnsemblePlot(PlotConfig):
         environ.setup_modelling()
         ds = environ.get_dataset()
         history = environ.get_history(subset='harvest')
+        optimiser = environ.get_optimiser()
+
         cm.create_group_mpl(
             self,
-            self.draw_figures(ds, history),
+            self.draw_figures(ds, history, optimiser),
             title=u'Waveform fits for the ensemble',
             section='fits',
             feather_icon='activity',
@@ -423,7 +425,7 @@ residuals for time domain comparisons (red filled), spectra of observed and
 synthetic traces for amplitude spectrum comparisons, or cross correlation
 traces.''')
 
-    def draw_figures(self, ds, history):
+    def draw_figures(self, ds, history, optimiser):
 
         color_parameter = self.color_parameter
         misfit_cutoff = self.misfit_cutoff
@@ -444,10 +446,8 @@ traces.''')
             target_index[target] = i, i+target.nmisfits
             i += target.nmisfits
 
-        gms = problem.combine_misfits(history.misfits)
-        isort = num.argsort(gms)[::-1]
-        gms = gms[isort]
-        models = history.models[isort, :]
+        gms = history.get_sorted_primary_misfits()[::-1]
+        models = history.get_sorted_primary_models()[::-1]
 
         if misfit_cutoff is not None:
             ibest = gms < misfit_cutoff
@@ -803,10 +803,14 @@ class FitsWaveformPlot(PlotConfig):
         mpl_init(fontsize=self.font_size)
         environ.setup_modelling()
         ds = environ.get_dataset()
+        optimiser = environ.get_optimiser()
+
+        environ.setup_modelling()
+
         history = environ.get_history(subset='harvest')
         cm.create_group_mpl(
             self,
-            self.draw_figures(ds, history),
+            self.draw_figures(ds, history, optimiser),
             title=u'Waveform fits for best model',
             section='fits',
             feather_icon='activity',
@@ -845,7 +849,7 @@ relative misfit contribution to the global misfit of the optimisation (bottom
 box, red).
 ''')
 
-    def draw_figures(self, ds, history):
+    def draw_figures(self, ds, history, optimiser):
 
         fontsize = self.font_size
         fontsize_title = self.font_size_title
@@ -864,18 +868,16 @@ box, red).
             target_index[target] = i, i+target.nmisfits
             i += target.nmisfits
 
-        gms = problem.combine_misfits(history.misfits)
-        isort = num.argsort(gms)
-        gms = gms[isort]
-        models = history.models[isort, :]
-        misfits = history.misfits[isort, :]
-
-        xbest = models[0, :]
+        xbest = history.get_best_model()
+        models = history.get_sorted_primary_models()
+        misfits = history.misfits[history.get_sorted_misfits_idx(chain=0), ...]
 
         ws = problem.get_target_weights()
 
         gcms = problem.combine_misfits(
-            misfits[:1, :, :], get_contributions=True)[0, :]
+            misfits[:1, :, :],
+            extra_correlated_weights=optimiser.get_correlated_weights(problem),
+            get_contributions=True)[0, :]
 
         w_max = num.nanmax(ws)
         gcm_max = num.nanmax(gcms)
@@ -1280,10 +1282,7 @@ location of the source.
         ws = problem.get_target_weights()
 
         if history:
-            gms = problem.combine_misfits(history.misfits)
-            isort = num.argsort(gms)
-            gms = gms[isort]
-            misfits = history.misfits[isort, :]
+            misfits = history.misfits[history.get_sorted_misfits_idx(), ...]
             gcms = problem.combine_misfits(
                 misfits[:1, :, :], get_contributions=True)[0, :]